+++ src/components/Pricing/PricingCalculator/Tabbed.tsx
@@ -1,188 +1,14 @@
 import React, { useEffect, useMemo, useState } from 'react'
 import Tooltip from 'components/Tooltip'
-import { IconInfo, IconLightBulb, IconX } from '@posthog/icons'
+import { IconInfo, IconLightBulb } from '@posthog/icons'
 import Toggle from 'components/Toggle'
-import { calculatePrice, formatUSD, pricingSliderLogic } from '../PricingSlider/pricingSliderLogic'
+import { calculatePrice, formatUSD } from '../PricingSlider/pricingSliderLogic'
 import { useStaticQuery } from 'gatsby'
 import { allProductsData } from '../Pricing'
-import { CallToAction } from 'components/CallToAction'
-import Link from 'components/Link'
 import useProducts from 'hooks/useProducts'
 import { LogSlider, inverseCurve, sliderCurve } from '../PricingSlider/Slider'
 import { PricingTiers } from '../Plans'
 import ProductAnalyticsTab from './Tabs/ProductAnalytics'
-
-const Modal = ({ onClose, isVisible }) => {
-    return (
-        <>
-            <div
-                className={`bg-accent-dark/50 fixed left-0 w-screen h-screen top-0 bg-opacity-40 flex justify-center items-center ${
-                    !isVisible ? 'hidden' : 'z-[1000000]'
-                }`}
-                onClick={() => onClose()}
-            ></div>
-            <div
-                className={`w-[500px] max-w-full h-[calc(100vh_-_123px)] md:h-screen z-[1000001] fixed overflow-y-auto top-0 flex flex-col bg-white dark:bg-accent-dark transition-all duration-500 ease-out
-          ${isVisible ? '!opacity-100 right-[0px]' : 'opacity-0 right-[-500px]'}`}
-            >
-                <div className="w-full h-fit flex justify-between p-4 md:px-8">
-                    <span className="font-bold text-xl">What are person profiles?</span>
-
-                    <button onClick={() => onClose()}>
-                        <IconX className="size-5" />
-                    </button>
-                </div>
-                <div className="px-4 pb-4 md:px-8 md:pb-8 [&_p]:text-[15px]">
-                    <p className="mb-2">
-                        Person profiles let you store detailed information about a user (like an email address, plan
-                        name, or custom properties). This enables a deeper level of insights that isn't available when
-                        analyzing anonymous traffic (like in Google Analytics).
-                    </p>
-                    <p className="mb-2">
-                        It also provides attribution data and lets you track users across devices and sessions.
-                    </p>
-                    <p className="mb-2">
-                        <strong>With person profiles, you can:</strong>
-                    </p>
-                    <ul className="[&_li]:text-[15px] pl-4 mb-2">
-                        <li>
-                            View detailed person info in the{' '}
-                            <Link href="https://us.posthog.com/persons" external>
-                                People tab
-                            </Link>
-                        </li>
-                        <li>
-                            Merge identities across <code className="text-sm">distinct_id</code> values (i.e. merge
-                            anonymous user and logged in users)
-                        </li>
-                        <li>Track UTMs and referrers for anonymous and identified users</li>
-                        <li>Target flags, experiments, and surveys by user properties</li>
-                        <li>Filter on user properties</li>
-                        <li>Create cohorts</li>
-                    </ul>
-
-                    <p className="text-[15px] mb-0">
-                        Person profiles are split out from standard event prices so you don't have to pay for them if
-                        you don't need them.
-                    </p>
-
-                    <div className="mb-8">
-                        <CallToAction href="/docs/data/persons" size="sm" className="mt-4" external>
-                            Read the docs
-                        </CallToAction>
-                    </div>
-
-                    <h3 className="text-xl border-b border-light dark:border-dark pb-1 mb-2">
-                        Do you need events with person profiles?
-                    </h3>
-
-                    <div className="space-y-2 mb-8">
-                        <details>
-                            <summary className="cursor-pointer">
-                                <div className="inline-flex items-center justify-between w-[95%]">
-                                    <strong>"I want to store properties on anonymous users"</strong>
-                                    <span>
-                                        <div className="border border-green rounded px-1 py-0.5 text-green font-bold uppercase text-sm">
-                                            Yes
-                                        </div>
-                                    </span>
-                                </div>
-                                <p className="text-sm text-primary/75 dark:text-primary-dark/75 pl-4 mb-0">
-                                    <span className="border border-red dark:border-yellow text-xs text-red dark:text-yellow font-semibold uppercase px-1 py-0.5 rounded-sm">
-                                        Most common
-                                    </span>{' '}
-                                    Read how to calculate this below
-                                </p>
-                            </summary>
-                            <div className="ml-4 my-2 border border-light dark:border-dark bg-accent dark:bg-accent-dark p-2 rounded [&_p]:text-sm">
-                                <p className="mb-2">
-                                    <strong className="text-green">
-                                        Yes, you'll use person profiles any time you provide person information to
-                                        store.
-                                    </strong>{' '}
-                                    This will allow you to store things like site preferences (eg. dark mode) and more
-                                    on non-identified users and still leverage all the benefits of person profiles.
-                                </p>
-                                <p className="mb-0">
-                                    It also allows you to merge anonymous and identified users, so you can track the
-                                    same user from the first time they visit your site through to when they sign up.
-                                </p>
-                            </div>
-                        </details>
-
-                        <details>
-                            <summary className="cursor-pointer">
-                                <div className="inline-flex items-center justify-between w-[95%]">
-                                    <strong>"I only want to track website visits (like Google Analytics)"</strong>
-                                    <span>
-                                        <div className="border border-red rounded px-1 py-0.5 text-red font-bold uppercase text-sm">
-                                            No
-                                        </div>
-                                    </span>
-                                </div>
-                                <p className="text-sm opacity-75 pl-4 mb-0">
-                                    Set <em>Enhanced persons</em> to <strong>0%</strong>
-                                </p>
-                            </summary>
-                            <div className="ml-4 my-2 border border-light dark:border-dark bg-accent dark:bg-accent-dark p-2 rounded [&_p]:text-sm">
-                                <p className="mb-0">
-                                    <strong>No, you don't need events with person profiles</strong> to track anonymous
-                                    users. You'll get basic attribution info (like UTM source) but it can't be used to
-                                    analyze things like top performing marketing channels as there's no way to know who
-                                    signed up.
-                                </p>
-                            </div>
-                        </details>
-
-                        <details>
-                            <summary className="cursor-pointer">
-                                <div className="inline-flex items-center justify-between w-[95%]">
-                                    <strong>"I only want to track logged in users"</strong>
-                                    <span>
-                                        <div className="border border-green rounded px-1 py-0.5 text-green font-bold uppercase text-sm">
-                                            Yes
-                                        </div>
-                                    </span>
-                                </div>
-                                <p className="text-sm opacity-75 pl-4 mb-0">
-                                    Set <em>Enhanced persons</em> to <strong>100%</strong>
-                                </p>
-                            </summary>
-                            <div className="ml-4 my-2 border border-light dark:border-dark bg-accent dark:bg-accent-dark p-2 rounded [&_p]:text-sm">
-                                <p className="mb-0">
-                                    <strong className="text-green">
-                                        Yes, you will use person profiles with any identified users
-                                    </strong>{' '}
-                                    so you can store information about the person (like an email address or internal
-                                    ID).
-                                </p>
-                            </div>
-                        </details>
-                    </div>
-
-                    <h3 className="text-xl border-b border-light dark:border-dark pb-1 mb-2">Calculating usage</h3>
-                    <p className="text-[15px] mb-2">
-                        If you already track website visits, the best way to gauge person profile usage is to determine
-                        what percentage of event volume is <em>anonymous traffic</em> vs. identified product users.
-                    </p>
-                    <p className="text-[15px] mb-2">
-                        If you currently use Google Analytics in conjunction with a product analytics tool, you can
-                        compare the traffic volume (page views) in Google Analytics and the event volume in your product
-                        anaytics solution to find what percentage of your event volume is anonymous.
-                    </p>
-                    <p className="text-[15px] mb-2">
-                        For example, if 75% of your total event volume (between website and product) is anonymous
-                        website traffic, you'll need person profiles for the remaining 25% of events.
-                    </p>
-                    <p className="text-[15px]">
-                        Note that you'll need to adjust this ratio if you want to store custom properties on anonymous
-                        users, as these require person profiles even if you don't know their email address yet.
-                    </p>
-                </div>
-            </div>
-        </>
-    )
-}
 
 const Addon = ({ type, name, description, plans, addons, setAddons, volume, inclusion_only }) => {
     const addon = addons.find((addon) => addon.type === type)
